requirements.txt
asgiref==3.8.1
Django==4.2.7
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
reportlab==4.4.0
sqlparse==0.5.3
whitenoise==6.6.0
